chore: break up "mitre/affiliation" structure
We had a lot going on in a single file, which made it a
bit difficult to follow when reading the code. This is
hopefully a start to making it easier.
